requirements.txt
streamlit
Pillow
torch
torchvision
ultralytics
opencv-python-headless